Practica Pre-Examen Entregables

+++ Ejercicios/src/main/java/org/velazquez/U5_herencia_interfaces/Practica_U5/Ex_Practica_20_21/Vehiculo.java
@@ -2,10 +2,10 @@ package org.velazquez.U5_herencia_interfaces.Practica_U5.Ex_Practica_20_21;
 
 public abstract class Vehiculo {
     private String matricula;
-    private float cargaMax;
+    private int cargaMax;
     private Conductor conductor;
 
-    public Vehiculo(String matricula, float cargaMax, Conductor conductor){
+    public Vehiculo(String matricula, int cargaMax, Conductor conductor){
         this.matricula=matricula;
         this.cargaMax=cargaMax;
         this.conductor=conductor;
